Refactoring: Convert function -> component
